backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
email-validator==2.1.1
httpx==0.27.0
# psycopg2-binary==2.9.9  # Только для PostgreSQL, закомментировано для SQLite
